+++ integration-tests/api/__tests__/claims/index.js
@@ -121,6 +121,83 @@ describe("Claims", () => {
     )
   })
 
+  test("creates a refund claim + return", async () => {
+    await adminSeeder(dbConnection)
+
+    const order = await createReturnableOrder(dbConnection)
+    const api = useApi()
+
+    const response = await api.post(
+      `/admin/orders/${order.id}/claims`,
+      {
+        type: "refund",
+        claim_items: [
+          {
+            item_id: "test-item",
+            reason: "missing_item",
+            quantity: 1,
+          },
+        ],
+        return_shipping: {
+          price: 0,
+        },
+      },
+      {
+        headers: {
+          authorization: "Bearer test_token",
+        },
+      }
+    )
+
+    const claim = response.data.order.claims[0]
+    const refund = response.data.order.refunds[0]
+    const returnOrder = response.data.order.returns[0]
+
+    expect(response.status).toEqual(200)
+    expect(claim.refund_amount).toEqual(1200)
+    expect(refund.amount).toEqual(1200)
+    expect(returnOrder.refund_amount).toEqual(1200)
+  })
+
+  test("creates a refund claim + return with a custom amount", async () => {
+    await adminSeeder(dbConnection)
+
+    const order = await createReturnableOrder(dbConnection)
+    const api = useApi()
+
+    const response = await api.post(
+      `/admin/orders/${order.id}/claims`,
+      {
+        type: "refund",
+        refund_amount: 500,
+        claim_items: [
+          {
+            item_id: "test-item",
+            reason: "missing_item",
+            quantity: 1,
+          },
+        ],
+        return_shipping: {
+          price: 0,
+        },
+      },
+      {
+        headers: {
+          authorization: "Bearer test_token",
+        },
+      }
+    )
+
+    const claim = response.data.order.claims[0]
+    const refund = response.data.order.refunds[0]
+    const returnOrder = response.data.order.returns[0]
+
+    expect(response.status).toEqual(200)
+    expect(claim.refund_amount).toEqual(500)
+    expect(refund.amount).toEqual(500)
+    expect(returnOrder.refund_amount).toEqual(500)
+  })
+
   test("creates a replace claim", async () => {
     await adminSeeder(dbConnection)