+++ integration-tests/api/helpers/admin-seeder.js
@@ -1,9 +1,8 @@
 const Scrypt = require("scrypt-kdf")
 const { User } = require("@medusajs/medusa")
-const { simpleSalesChannelFactory } = require("../factories")
 
-module.exports = async (connection, data = {}) => {
-  const manager = connection.manager
+module.exports = async (dataSource, data = {}) => {
+  const manager = dataSource.manager
 
   const buf = await Scrypt.kdf("secret_password", { logN: 1, r: 1, p: 1 })
   const password_hash = buf.toString("base64")