+++ packages/medusa/src/services/__tests__/cart.js
@@ -65,7 +65,7 @@ describe("CartService", () => {
   describe("retrieve", () => {
     let result
     const cartRepository = MockRepository({
-      findOneWithRelations: () =>
+      findOne: () =>
         Promise.resolve({ id: IdMap.getId("emptyCart") }),
     })
     beforeAll(async () => {
@@ -82,9 +82,8 @@ describe("CartService", () => {
     })
 
     it("calls cart model functions", () => {
-      expect(cartRepository.findOneWithRelations).toHaveBeenCalledTimes(1)
-      expect(cartRepository.findOneWithRelations).toHaveBeenCalledWith(
-        undefined,
+      expect(cartRepository.findOne).toHaveBeenCalledTimes(1)
+      expect(cartRepository.findOne).toHaveBeenCalledWith(
         {
           where: { id: IdMap.getId("emptyCart") },
           select: undefined,
@@ -129,7 +128,7 @@ describe("CartService", () => {
       )
 
       expect(cartRepository.findOne).toBeCalledTimes(1)
-      expect(cartRepository.findOne).toBeCalledWith(id)
+      expect(cartRepository.findOne).toBeCalledWith({ where: { id } })
 
       expect(cartRepository.save).toBeCalledTimes(1)
       expect(cartRepository.save).toBeCalledWith({
@@ -346,7 +345,7 @@ describe("CartService", () => {
     }
 
     const cartRepository = MockRepository({
-      findOneWithRelations: (rels, q) => {
+      findOne: (q) => {
         if (q.where.id === IdMap.getId("cartWithLine")) {
           return Promise.resolve({
             id: IdMap.getId("cartWithLine"),
@@ -587,7 +586,7 @@ describe("CartService", () => {
     }
 
     const cartRepository = MockRepository({
-      findOneWithRelations: (rels, q) => {
+      findOne: (q) => {
         if (q.where.id === IdMap.getId("cartWithLine")) {
           return Promise.resolve({
             id: IdMap.getId("cartWithLine"),
@@ -673,7 +672,7 @@ describe("CartService", () => {
       },
     }
     const cartRepository = MockRepository({
-      findOneWithRelations: (rels, q) => {
+      findOne: (q) => {
         if (q.where.id === IdMap.getId("withShipping")) {
           return Promise.resolve({
             shipping_methods: [
@@ -816,7 +815,7 @@ describe("CartService", () => {
 
   describe("update", () => {
     const cartRepository = MockRepository({
-      findOneWithRelations: (rels, q) => {
+      findOne: (q) => {
         if (q.where.id === "withpays") {
           return Promise.resolve({
             payment_sessions: [
@@ -847,24 +846,31 @@ describe("CartService", () => {
       cartService.setPaymentSessions = jest.fn()
       await cartService.update("withpays", {})
 
-      expect(cartRepository.findOneWithRelations).toHaveBeenCalledWith(
-        expect.arrayContaining([
-          "items",
-          "shipping_methods",
-          "shipping_address",
-          "billing_address",
-          "gift_cards",
-          "customer",
-          "region",
-          "payment_sessions",
-          "region.countries",
-          "discounts",
-          "discounts.rule",
-          "discounts.regions",
-        ]),
-        {
-          where: { id: "withpays" },
-        }
+      expect(cartRepository.findOne).toHaveBeenCalledWith(
+        expect.objectContaining({
+          relations: {
+            billing_address: true,
+            customer: true,
+            discounts: {
+              regions: true,
+              rule: true
+            },
+            gift_cards: true,
+            items: {
+              variant: {
+                product: true
+              }
+            },
+            payment_sessions: true,
+            region: { countries: true },
+            shipping_address: true,
+            shipping_methods: true
+          },
+          select: undefined,
+          where: {
+            id: "withpays"
+          }
+        }),
       )
     })
   })
@@ -904,7 +910,7 @@ describe("CartService", () => {
     }
 
     const cartRepository = MockRepository({
-      findOneWithRelations: (rels, q) => {
+      findOne: (q) => {
         if (q.where.id === IdMap.getId("cannot")) {
           return Promise.resolve({
             items: [
@@ -1016,7 +1022,7 @@ describe("CartService", () => {
       },
     }
     const cartRepository = MockRepository({
-      findOneWithRelations: () => Promise.resolve({}),
+      findOne: () => Promise.resolve({}),
     })
     const cartService = new CartService({
       manager: MockManager,
@@ -1088,7 +1094,7 @@ describe("CartService", () => {
 
   describe("updateBillingAddress", () => {
     const cartRepository = MockRepository({
-      findOneWithRelations: () =>
+      findOne: () =>
         Promise.resolve({
           region: { countries: [{ iso_2: "us" }] },
         }),
@@ -1150,7 +1156,7 @@ describe("CartService", () => {
 
   describe("updateShippingAddress", () => {
     const cartRepository = MockRepository({
-      findOneWithRelations: () =>
+      findOne: () =>
         Promise.resolve({
           region: { countries: [{ iso_2: "us" }] },
         }),
@@ -1231,6 +1237,16 @@ describe("CartService", () => {
     const lineItemService = {
       update: jest.fn((r) => r),
       delete: jest.fn(),
+      retrieve: jest.fn().mockImplementation((lineItemId) => {
+        if (lineItemId === IdMap.getId("existing")) {
+          return Promise.resolve({
+            id: lineItemId,
+          })
+        }
+        return Promise.resolve({
+          id: lineItemId,
+        })
+      }),
       withTransaction: function () {
         return this
       },
@@ -1256,7 +1272,7 @@ describe("CartService", () => {
       },
     }
     const cartRepository = MockRepository({
-      findOneWithRelations: () =>
+      findOne: () =>
         Promise.resolve({
           items: [
             {
@@ -1361,7 +1377,9 @@ describe("CartService", () => {
           shipping_address: {
             country_code: "us",
           },
-          items: [IdMap.getId("testitem")],
+          items: [{
+            id: IdMap.getId("testitem")
+          }],
           payment_session: null,
           payment_sessions: [],
           gift_cards: [],
@@ -1378,7 +1396,7 @@ describe("CartService", () => {
 
   describe("setPaymentSession", () => {
     const cartRepository = MockRepository({
-      findOneWithRelations: (rels, q) => {
+      findOne: (q) => {
         if (q.where.id === IdMap.getId("cartWithLine")) {
           return Promise.resolve({
             total: 100,
@@ -1586,7 +1604,7 @@ describe("CartService", () => {
     }
 
     const cartRepository = MockRepository({
-      findOneWithRelations: (rels, q) => {
+      findOne: (q) => {
         if (q.where.id === IdMap.getId("cart-to-filter")) {
           return Promise.resolve(cart3)
         }
@@ -1792,7 +1810,7 @@ describe("CartService", () => {
     const cartWithCustomSO = buildCart("cart-with-custom-so")
 
     const cartRepository = MockRepository({
-      findOneWithRelations: (rels, q) => {
+      findOne: (q) => {
         switch (q.where.id) {
           case IdMap.getId("lines"):
             return Promise.resolve(cart3)
@@ -1991,7 +2009,7 @@ describe("CartService", () => {
     }
 
     const cartRepository = MockRepository({
-      findOneWithRelations: (rels, q) => {
+      findOne: (q) => {
         if (q.where.id === IdMap.getId("with-d")) {
           return Promise.resolve({
             id: IdMap.getId("cart"),
@@ -2480,7 +2498,7 @@ describe("CartService", () => {
 
   describe("removeDiscount", () => {
     const cartRepository = MockRepository({
-      findOneWithRelations: (rels, q) => {
+      findOne: (q) => {
         return Promise.resolve({
           id: IdMap.getId("cart"),
           discounts: [