+++ packages/medusa/src/services/__tests__/swap.ts
@@ -1,7 +1,11 @@
 import { IdMap, MockManager, MockRepository } from "medusa-test-utils"
 import SwapService from "../swap"
-import { ProductVariantInventoryServiceMock } from "../__mocks__/product-variant-inventory"
-import { LineItemAdjustmentServiceMock } from "../__mocks__/line-item-adjustment"
+import {
+  ProductVariantInventoryServiceMock
+} from "../__mocks__/product-variant-inventory"
+import {
+  LineItemAdjustmentServiceMock
+} from "../__mocks__/line-item-adjustment"
 import {
   CustomShippingOptionService,
   EventBusService,
@@ -22,7 +26,7 @@ import LineItemAdjustmentService from "../line-item-adjustment"
 /* ******************** DEFAULT REPOSITORY MOCKS ******************** */
 
 const swapRepo = MockRepository({
-  findOneWithRelations: (existing) => Promise.resolve(existing),
+  findOne: (existing) => Promise.resolve(existing),
   create: jest.fn().mockImplementation((data) => {
     return Object.assign(new Swap(), data)
   }),
@@ -65,6 +69,7 @@ const lineItemService = {
   create: jest.fn().mockImplementation((d) => Promise.resolve(d)),
   update: jest.fn().mockImplementation((d) => Promise.resolve(d)),
   retrieve: () => Promise.resolve({}),
+  list: () => Promise.resolve([]),
   createReturnLines: jest.fn(() => Promise.resolve()),
   withTransaction: function () {
     return this
@@ -225,7 +230,7 @@ describe("SwapService", () => {
       } as unknown as CartService
 
       const swapRepo = MockRepository({
-        findOneWithRelations: () => Promise.resolve(existing),
+        findOne: () => Promise.resolve(existing),
       })
 
       const customShippingOptionService = {
@@ -240,6 +245,7 @@ describe("SwapService", () => {
         create: jest.fn().mockImplementation((d) => Promise.resolve(d)),
         update: jest.fn().mockImplementation((d) => Promise.resolve(d)),
         retrieve: () => Promise.resolve({}),
+        list: () => Promise.resolve([]),
         createReturnLines: jest.fn(() => Promise.resolve()),
         withTransaction: function () {
           return this
@@ -263,28 +269,37 @@ describe("SwapService", () => {
           { option_id: "test-option", price: 10 },
         ])
 
-        expect(swapRepo.findOneWithRelations).toHaveBeenCalledTimes(1)
-        expect(swapRepo.findOneWithRelations).toHaveBeenCalledWith(
-          [
-            "order",
-            "order.items",
-            "order.swaps",
-            "order.swaps.additional_items",
-            "order.discounts",
-            "order.discounts.rule",
-            "order.claims",
-            "order.claims.additional_items",
-            "additional_items",
-            "additional_items.variant",
-            "return_order",
-            "return_order.items",
-            "return_order.shipping_method",
-            "return_order.shipping_method.tax_lines",
-          ],
-          {
-            where: { id: IdMap.getId("swap-1") },
-          }
-        )
+        expect(swapRepo.findOne).toHaveBeenCalledTimes(1)
+        expect(swapRepo.findOne).toHaveBeenCalledWith({
+          relations: {
+            additional_items: {
+              variant: true
+            },
+            order: {
+              claims: {
+                additional_items: true
+              },
+              discounts: {
+                rule: true
+              },
+              items: {
+                variant: {
+                  product: true
+                }
+              },
+              swaps: {
+                additional_items: true
+              }
+            },
+            return_order: {
+              items: true,
+              shipping_method: {
+                tax_lines: true
+              }
+            }
+          },
+          where: { id: IdMap.getId("swap-1") }
+        })
 
         expect(lineItemService.createReturnLines).toHaveBeenCalledTimes(1)
         expect(lineItemService.createReturnLines).toHaveBeenCalledWith(
@@ -352,7 +367,7 @@ describe("SwapService", () => {
       }
 
       const swapRepo = MockRepository({
-        findOneWithRelations: (_, query) => {
+        findOne: (query) => {
           switch (query.where.id) {
             case IdMap.getId("canceled"):
               return Promise.resolve({
@@ -400,7 +415,7 @@ describe("SwapService", () => {
       const lineItemService = {
         generate: jest
           .fn()
-          .mockImplementation((variantId, regionId, quantity) => {
+          .mockImplementation(({ variantId, quantity }) => {
             return {
               unit_price: 100,
               variant_id: variantId,
@@ -408,6 +423,7 @@ describe("SwapService", () => {
             }
           }),
         retrieve: () => Promise.resolve({}),
+        list: () => Promise.resolve([]),
         withTransaction: function () {
           return this
         },
@@ -441,11 +457,13 @@ describe("SwapService", () => {
         )
 
         expect(lineItemService.generate).toHaveBeenCalledTimes(1)
-        expect(lineItemService.generate).toHaveBeenCalledWith(
-          IdMap.getId("new-variant"),
-          IdMap.getId("region"),
-          1
-        )
+        expect(lineItemService.generate).toHaveBeenCalledWith({
+          quantity: 1,
+          variantId: IdMap.getId("new-variant")
+        }, {
+          "cart": undefined,
+          region_id: IdMap.getId("region")
+        })
       })
 
       it("creates swap", async () => {
@@ -538,13 +556,14 @@ describe("SwapService", () => {
       const lineItemService = {
         update: jest.fn(),
         retrieve: () => Promise.resolve({}),
+        list: () => Promise.resolve([]),
         withTransaction: function () {
           return this
         },
       } as unknown as LineItemService
 
       const swapRepo = MockRepository({
-        findOneWithRelations: () => Promise.resolve({ ...existing }),
+        findOne: () => Promise.resolve({ ...existing }),
       })
       const swapService = new SwapService({
         ...defaultProps,
@@ -586,14 +605,14 @@ describe("SwapService", () => {
             shipping_methods: existing.shipping_methods,
           },
           [{ item_id: "1234", quantity: 2 }],
-          { swap_id: IdMap.getId("swap"), metadata: {} }
+          { swap_id: IdMap.getId("swap"), metadata: {} },
         )
       })
     })
 
     describe("failure", () => {
       const swapRepo = MockRepository({
-        findOneWithRelations: () =>
+        findOne: () =>
           Promise.resolve({
             canceled_at: new Date(),
           }),
@@ -615,7 +634,7 @@ describe("SwapService", () => {
 
   describe("cancelFulfillment", () => {
     const swapRepo = MockRepository({
-      findOneWithRelations: () => Promise.resolve({}),
+      findOne: () => Promise.resolve({}),
       save: (f) => Promise.resolve(f),
     })
 
@@ -736,7 +755,7 @@ describe("SwapService", () => {
       } as unknown as CartService
 
       const swapRepo = MockRepository({
-        findOneWithRelations: () => Promise.resolve(existing),
+        findOne: () => Promise.resolve(existing),
       })
 
       const swapService = new SwapService({
@@ -779,7 +798,7 @@ describe("SwapService", () => {
 
     describe("failure", () => {
       const swapRepo = MockRepository({
-        findOneWithRelations: () =>
+        findOne: () =>
           Promise.resolve({ canceled_at: new Date() }),
       })
 
@@ -882,7 +901,7 @@ describe("SwapService", () => {
         cart) as unknown as CartService["retrieveWithTotals"]
 
       const swapRepo = MockRepository({
-        findOneWithRelations: () => Promise.resolve(existing),
+        findOne: () => Promise.resolve(existing),
       })
 
       const swapService = new SwapService({
@@ -934,7 +953,7 @@ describe("SwapService", () => {
       }
 
       const swapRepo = MockRepository({
-        findOneWithRelations: (rels, q) => {
+        findOne: (q) => {
           switch (q.where.id) {
             case IdMap.getId("canceled"):
               return Promise.resolve({ canceled_at: new Date() })
@@ -1001,7 +1020,7 @@ describe("SwapService", () => {
       })
 
       const swapRepo = MockRepository({
-        findOneWithRelations: (rels, q) => {
+        findOne: (q) => {
           switch (q.where.id) {
             case "refund":
               return Promise.resolve(existing(-1, false))
@@ -1123,7 +1142,7 @@ describe("SwapService", () => {
       })
 
       const swapRepo = MockRepository({
-        findOneWithRelations: (rels, q) => {
+        findOne: (q) => {
           switch (q.where.id) {
             case "requested":
               return Promise.resolve({
@@ -1180,7 +1199,7 @@ describe("SwapService", () => {
     const fulfillment = { canceled_at: now }
 
     const swapRepo = MockRepository({
-      findOneWithRelations: (_, q) => {
+      findOne: (q) => {
         const swap: any = {
           payment: { ...payment },
           return_order: { ...return_order },