+++ packages/medusa/src/services/cart.ts
@@ -203,7 +203,7 @@ class CartService extends TransactionBaseService {
     config: FindConfig<Cart> = {}
   ): Promise<Cart[]> {
     const manager = this.manager_
-    const cartRepo = manager.getCustomRepository(this.cartRepository_)
+    const cartRepo = manager.withRepository(this.cartRepository_)
 
     const query = buildQuery(selector, config)
     return await cartRepo.find(query)
@@ -235,18 +235,14 @@ class CartService extends TransactionBaseService {
     }
 
     const manager = this.manager_
-    const cartRepo = manager.getCustomRepository(this.cartRepository_)
-
+    const cartRepo = manager.withRepository(this.cartRepository_)
     const query = buildQuery({ id: cartId }, options)
 
     if ((options.select || []).length === 0) {
       query.select = undefined
     }
 
-    const queryRelations = query.relations
-    query.relations = undefined
-
-    const raw = await cartRepo.findOneWithRelations(queryRelations, query)
+    const raw = await cartRepo.findOne(query)
 
     if (!raw) {
       throw new MedusaError(
@@ -271,22 +267,19 @@ class CartService extends TransactionBaseService {
     totalsConfig: TotalsConfig = {}
   ): Promise<Cart> {
     const manager = this.manager_
-    const cartRepo = manager.getCustomRepository(this.cartRepository_)
+    const cartRepo = manager.withRepository(this.cartRepository_)
 
     const { select, relations, totalsToSelect } =
       this.transformQueryForTotals_(options)
 
-    const query = buildQuery({ id: cartId }, { ...options, select, relations })
+    const query = buildQuery({ id: cartId }, {
+      ...options,
+      select,
+      relations,
+    } as FindConfig<Cart>)
 
-    if (relations && relations.length > 0) {
-      query.relations = relations
-    }
+    const raw = await cartRepo.findOne(query)
 
-    query.select = select?.length ? select : undefined
-
-    const queryRelations = query.relations
-    query.relations = undefined
-    const raw = await cartRepo.findOneWithRelations(queryRelations, query)
     if (!raw) {
       throw new MedusaError(
         MedusaError.Types.NOT_FOUND,
@@ -320,10 +313,8 @@ class CartService extends TransactionBaseService {
   async create(data: CartCreateProps): Promise<Cart | never> {
     return await this.atomicPhase_(
       async (transactionManager: EntityManager) => {
-        const cartRepo = transactionManager.getCustomRepository(
-          this.cartRepository_
-        )
-        const addressRepo = transactionManager.getCustomRepository(
+        const cartRepo = transactionManager.withRepository(this.cartRepository_)
+        const addressRepo = transactionManager.withRepository(
           this.addressRepository_
         )
 
@@ -392,7 +383,9 @@ class CartService extends TransactionBaseService {
             rawCart.shipping_address = data.shipping_address
           }
           if (data.shipping_address_id) {
-            const addr = await addressRepo.findOne(data.shipping_address_id)
+            const addr = await addressRepo.findOne({
+              where: { id: data.shipping_address_id },
+            })
             if (
               addr?.country_code &&
               !regCountries.includes(addr.country_code)
@@ -416,7 +409,9 @@ class CartService extends TransactionBaseService {
           rawCart.billing_address = data.billing_address
         }
         if (data.billing_address_id) {
-          const addr = await addressRepo.findOne(data.billing_address_id)
+          const addr = await addressRepo.findOne({
+            where: { id: data.billing_address_id },
+          })
           if (addr?.country_code && !regCountries.includes(addr.country_code)) {
             throw new MedusaError(
               MedusaError.Types.NOT_ALLOWED,
@@ -509,7 +504,7 @@ class CartService extends TransactionBaseService {
             .deleteShippingMethods(cart.shipping_methods)
         }
 
-        const lineItemRepository = transactionManager.getCustomRepository(
+        const lineItemRepository = transactionManager.withRepository(
           this.lineItemRepository_
         )
         await lineItemRepository.update(
@@ -526,7 +521,14 @@ class CartService extends TransactionBaseService {
           .delete(lineItem.id)
 
         const result = await this.retrieve(cartId, {
-          relations: ["items", "discounts", "discounts.rule", "region"],
+          relations: [
+            "items",
+            "items.variant",
+            "items.variant.product",
+            "discounts",
+            "discounts.rule",
+            "region"
+          ],
         })
 
         await this.refreshAdjustments_(result)
@@ -720,7 +722,7 @@ class CartService extends TransactionBaseService {
           })
 
         cart = await this.retrieve(cart.id, {
-          relations: ["items", "discounts", "discounts.rule", "region"],
+          relations: ["items", "items.variant", "items.variant.product", "discounts", "discounts.rule", "region"],
         })
 
         await this.refreshAdjustments_(cart)
@@ -893,7 +895,14 @@ class CartService extends TransactionBaseService {
           })
 
         cart = await this.retrieve(cart.id, {
-          relations: ["items", "discounts", "discounts.rule", "region"],
+          relations: [
+            "items",
+            "items.variant",
+            "items.variant.product",
+            "discounts",
+            "discounts.rule",
+            "region"
+          ],
         })
 
         await this.refreshAdjustments_(cart)
@@ -965,7 +974,14 @@ class CartService extends TransactionBaseService {
           .update(lineItemId, lineItemUpdate)
 
         const updatedCart = await this.retrieve(cartId, {
-          relations: ["items", "discounts", "discounts.rule", "region"],
+          relations: [
+            "items",
+            "items.variant",
+            "items.variant.product",
+            "discounts",
+            "discounts.rule",
+            "region"
+          ],
         })
 
         await this.refreshAdjustments_(updatedCart)
@@ -996,7 +1012,7 @@ class CartService extends TransactionBaseService {
     const transactionManager = this.transactionManager_ ?? this.manager_
 
     if (cart.shipping_methods?.length) {
-      const shippingMethodRepository = transactionManager.getCustomRepository(
+      const shippingMethodRepository = transactionManager.withRepository(
         this.shippingMethodRepository_
       )
 
@@ -1035,11 +1051,11 @@ class CartService extends TransactionBaseService {
   async update(cartId: string, data: CartUpdateProps): Promise<Cart> {
     return await this.atomicPhase_(
       async (transactionManager: EntityManager) => {
-        const cartRepo = transactionManager.getCustomRepository(
-          this.cartRepository_
-        )
+        const cartRepo = transactionManager.withRepository(this.cartRepository_)
         const relations = [
           "items",
+          "items.variant",
+          "items.variant.product",
           "shipping_methods",
           "shipping_address",
           "billing_address",
@@ -1092,7 +1108,7 @@ class CartService extends TransactionBaseService {
           await this.setRegion_(cart, data.region_id, countryCode)
         }
 
-        const addrRepo = transactionManager.getCustomRepository(
+        const addrRepo = transactionManager.withRepository(
           this.addressRepository_
         )
 
@@ -1292,7 +1308,7 @@ class CartService extends TransactionBaseService {
   protected async updateBillingAddress_(
     cart: Cart,
     addressOrId: AddressPayload | Partial<Address> | string,
-    addrRepo: AddressRepository
+    addrRepo: typeof AddressRepository
   ): Promise<void> {
     let address: Address
     if (typeof addressOrId === `string`) {
@@ -1330,7 +1346,7 @@ class CartService extends TransactionBaseService {
   protected async updateShippingAddress_(
     cart: Cart,
     addressOrId: AddressPayload | Partial<Address> | string,
-    addrRepo: AddressRepository
+    addrRepo: typeof AddressRepository
   ): Promise<void> {
     let address: Address
 
@@ -1505,6 +1521,8 @@ class CartService extends TransactionBaseService {
         const cart = await this.retrieve(cartId, {
           relations: [
             "items",
+            "items.variant",
+            "items.variant.product",
             "region",
             "discounts",
             "discounts.rule",
@@ -1525,9 +1543,7 @@ class CartService extends TransactionBaseService {
           (discount) => discount.code !== discountCode
         )
 
-        const cartRepo = transactionManager.getCustomRepository(
-          this.cartRepository_
-        )
+        const cartRepo = transactionManager.withRepository(this.cartRepository_)
         const updatedCart = await cartRepo.save(cart)
 
         await this.refreshAdjustments_(updatedCart)
@@ -1597,12 +1613,12 @@ class CartService extends TransactionBaseService {
   ): Promise<Cart> {
     return await this.atomicPhase_(
       async (transactionManager: EntityManager) => {
-        const cartRepository = transactionManager.getCustomRepository(
+        const cartRepository = transactionManager.withRepository(
           this.cartRepository_
         )
 
         const cart = await this.retrieveWithTotals(cartId, {
-          relations: ["payment_sessions"],
+          relations: ["payment_sessions", "items", "items.variant", "items.variant.product"],
         })
 
         // If cart total is 0, we don't perform anything payment related
@@ -1666,12 +1682,15 @@ class CartService extends TransactionBaseService {
   async setPaymentSession(cartId: string, providerId: string): Promise<void> {
     return await this.atomicPhase_(
       async (transactionManager: EntityManager) => {
-        const psRepo = transactionManager.getCustomRepository(
+        const psRepo = transactionManager.withRepository(
           this.paymentSessionRepository_
         )
 
         const cart = await this.retrieveWithTotals(cartId, {
           relations: [
+            "items",
+            "items.variant",
+            "items.variant.product",
             "customer",
             "region",
             "region.payment_providers",
@@ -1698,7 +1717,7 @@ class CartService extends TransactionBaseService {
           currentlySelectedSession &&
           currentlySelectedSession.provider_id !== providerId
         ) {
-          const psRepo = transactionManager.getCustomRepository(
+          const psRepo = transactionManager.withRepository(
             this.paymentSessionRepository_
           )
 
@@ -1781,7 +1800,7 @@ class CartService extends TransactionBaseService {
   async setPaymentSessions(cartOrCartId: Cart | string): Promise<void> {
     return await this.atomicPhase_(
       async (transactionManager: EntityManager) => {
-        const psRepo = transactionManager.getCustomRepository(
+        const psRepo = transactionManager.withRepository(
           this.paymentSessionRepository_
         )
 
@@ -1796,6 +1815,8 @@ class CartService extends TransactionBaseService {
           {
             relations: [
               "items",
+              "items.variant",
+              "items.variant.product",
               "items.adjustments",
               "discounts",
               "discounts.rule",
@@ -1966,9 +1987,7 @@ class CartService extends TransactionBaseService {
           relations: ["payment_sessions"],
         })
 
-        const cartRepo = transactionManager.getCustomRepository(
-          this.cartRepository_
-        )
+        const cartRepo = transactionManager.withRepository(this.cartRepository_)
 
         if (cart.payment_sessions) {
           const paymentSession = cart.payment_sessions.find(
@@ -1979,7 +1998,7 @@ class CartService extends TransactionBaseService {
             ({ provider_id }) => provider_id !== providerId
           )
 
-          const psRepo = transactionManager.getCustomRepository(
+          const psRepo = transactionManager.withRepository(
             this.paymentSessionRepository_
           )
 
@@ -1989,7 +2008,7 @@ class CartService extends TransactionBaseService {
                 .withTransaction(transactionManager)
                 .deleteSession(paymentSession)
             } else {
-              await psRepo.delete(paymentSession)
+              await psRepo.delete({ id: paymentSession.id })
             }
           }
         }
@@ -2037,7 +2056,7 @@ class CartService extends TransactionBaseService {
                   provider_id: providerId,
                 })
             } else {
-              const psRepo = transactionManager.getCustomRepository(
+              const psRepo = transactionManager.withRepository(
                 this.paymentSessionRepository_
               )
               await psRepo.update(paymentSession.id, {
@@ -2079,6 +2098,7 @@ class CartService extends TransactionBaseService {
             "items",
             "items.variant",
             "items.variant.product",
+            "payment_sessions",
           ],
         })
 
@@ -2230,9 +2250,13 @@ class CartService extends TransactionBaseService {
               availablePrice !== undefined &&
               availablePrice.calculatedPrice !== null
             ) {
-              return await lineItemServiceTx.update(item.id, {
+              await lineItemServiceTx.update(item.id, {
                 has_shipping: false,
                 unit_price: availablePrice.calculatedPrice,
+              })
+
+              return await lineItemServiceTx.retrieve(item.id, {
+                relations: ["variant", "variant.product"],
               })
             }
 
@@ -2274,9 +2298,7 @@ class CartService extends TransactionBaseService {
     cart.region = region
     cart.region_id = region.id
 
-    const addrRepo = transactionManager.getCustomRepository(
-      this.addressRepository_
-    )
+    const addrRepo = transactionManager.withRepository(this.addressRepository_)
     /*
      * When changing the region you are changing the set of countries that your
      * cart can be shipped to so we need to make sure that the current shipping
@@ -2368,7 +2390,7 @@ class CartService extends TransactionBaseService {
     cart.gift_cards = []
 
     if (cart.payment_sessions && cart.payment_sessions.length) {
-      const paymentSessionRepo = transactionManager.getCustomRepository(
+      const paymentSessionRepo = transactionManager.withRepository(
         this.paymentSessionRepository_
       )
       await paymentSessionRepo.delete({
@@ -2392,6 +2414,8 @@ class CartService extends TransactionBaseService {
         const cart = await this.retrieve(cartId, {
           relations: [
             "items",
+            "items.variant",
+            "items.variant.product",
             "discounts",
             "discounts.rule",
             "payment_sessions",
@@ -2412,9 +2436,7 @@ class CartService extends TransactionBaseService {
           )
         }
 
-        const cartRepo = transactionManager.getCustomRepository(
-          this.cartRepository_
-        )
+        const cartRepo = transactionManager.withRepository(this.cartRepository_)
         return cartRepo.remove(cart)
       }
     )
@@ -2436,9 +2458,7 @@ class CartService extends TransactionBaseService {
   ): Promise<Cart> {
     return await this.atomicPhase_(
       async (transactionManager: EntityManager) => {
-        const cartRepo = transactionManager.getCustomRepository(
-          this.cartRepository_
-        )
+        const cartRepo = transactionManager.withRepository(this.cartRepository_)
 
         if (typeof key !== "string") {
           throw new MedusaError(
@@ -2447,7 +2467,7 @@ class CartService extends TransactionBaseService {
           )
         }
 
-        const cart = await cartRepo.findOne(cartId)
+        const cart = await cartRepo.findOne({ where: { id: cartId } })
         if (!cart) {
           throw new MedusaError(
             MedusaError.Types.NOT_FOUND,
@@ -2483,6 +2503,8 @@ class CartService extends TransactionBaseService {
                 "discounts.rule",
                 "gift_cards",
                 "items",
+                "items.variant",
+                "items.variant.product",
                 "items.adjustments",
                 "region",
                 "region.tax_rates",
@@ -2749,6 +2771,8 @@ class CartService extends TransactionBaseService {
     const relationSet = new Set(config.relations)
 
     relationSet.add("items")
+    relationSet.add("items.variant")
+    relationSet.add("items.variant.product")
     relationSet.add("items.tax_lines")
     relationSet.add("items.adjustments")
     relationSet.add("gift_cards")