+++ packages/medusa/src/services/claim-item.ts
@@ -48,9 +48,7 @@ class ClaimItemService extends TransactionBaseService {
 
   async create(data: CreateClaimItemInput): Promise<ClaimItem> {
     return await this.atomicPhase_(async (manager) => {
-      const ciRepo: ClaimItemRepository = manager.getCustomRepository(
-        this.claimItemRepository_
-      )
+      const ciRepo = manager.withRepository(this.claimItemRepository_)
 
       const { item_id, reason, quantity, tags, images, ...rest } = data
 
@@ -86,9 +84,7 @@ class ClaimItemService extends TransactionBaseService {
 
       let tagsToAdd: ClaimTag[] = []
       if (tags && tags.length) {
-        const claimTagRepo = manager.getCustomRepository(
-          this.claimTagRepository_
-        )
+        const claimTagRepo = manager.withRepository(this.claimTagRepository_)
         tagsToAdd = await Promise.all(
           tags.map(async (t) => {
             const normalized = t.trim().toLowerCase()
@@ -105,9 +101,7 @@ class ClaimItemService extends TransactionBaseService {
 
       let imagesToAdd: ClaimImage[] = []
       if (images && images.length) {
-        const claimImgRepo = manager.getCustomRepository(
-          this.claimImageRepository_
-        )
+        const claimImgRepo = manager.withRepository(this.claimImageRepository_)
         imagesToAdd = images.map((url) => {
           return claimImgRepo.create({ url })
         })
@@ -138,7 +132,7 @@ class ClaimItemService extends TransactionBaseService {
 
   async update(id, data): Promise<ClaimItem> {
     return this.atomicPhase_(async (manager) => {
-      const ciRepo = manager.getCustomRepository(this.claimItemRepository_)
+      const ciRepo = manager.withRepository(this.claimItemRepository_)
       const item = await this.retrieve(id, { relations: ["images", "tags"] })
 
       const { tags, images, reason, note, metadata } = data
@@ -157,9 +151,7 @@ class ClaimItemService extends TransactionBaseService {
 
       if (tags) {
         item.tags = []
-        const claimTagRepo = manager.getCustomRepository(
-          this.claimTagRepository_
-        )
+        const claimTagRepo = manager.withRepository(this.claimTagRepository_)
         for (const t of tags) {
           if (t.id) {
             item.tags.push(t)
@@ -180,9 +172,7 @@ class ClaimItemService extends TransactionBaseService {
       }
 
       if (images) {
-        const claimImgRepo = manager.getCustomRepository(
-          this.claimImageRepository_
-        )
+        const claimImgRepo = manager.withRepository(this.claimImageRepository_)
         const ids = images.map((i) => i.id)
         for (const i of item.images) {
           if (!ids.includes(i.id)) {
@@ -226,7 +216,7 @@ class ClaimItemService extends TransactionBaseService {
       order: { created_at: "DESC" },
     }
   ): Promise<ClaimItem[]> {
-    const ciRepo = this.manager_.getCustomRepository(this.claimItemRepository_)
+    const ciRepo = this.manager_.withRepository(this.claimItemRepository_)
     const query = buildQuery(selector, config)
     return ciRepo.find(query)
   }
@@ -248,7 +238,7 @@ class ClaimItemService extends TransactionBaseService {
       )
     }
 
-    const claimItemRepo = this.manager_.getCustomRepository(
+    const claimItemRepo = this.manager_.withRepository(
       this.claimItemRepository_
     )
     const query = buildQuery({ id: claimItemId }, config)